refactored parts of the framework into separate files
