+++ src/SeeSign.jl
@@ -4,6 +4,9 @@ include("higher_order.jl")
 include("changed.jl")
 include("tracked.jl")
 include("depnet.jl")
+include("physical.jl")
+include("transitions.jl")
+include("generators.jl")
 include("framework.jl")
 include("sim.jl")
 include("regex_tuples.jl")
