+++ src/framework.jl
@@ -4,264 +4,8 @@ using CompetingClocks: SSA, CombinedNextReaction, enable!, disable!, next, Xoshi
 using Distributions
 
 
-####### The physical state is transactional.
-
-export PhysicalState, changed, wasread, resetread, accept
-abstract type PhysicalState end
-
-"""
-    isconsistent(physical_state)
-
-A simulation in debug mode will assert `isconsistent(physical_state)` is true.
-Override this to verify the physical state of your simulation.
-"""
-isconsistent(::PhysicalState) = true
-
-"""
-Iterate over all tracked vectors in the physical state.
-"""
-function over_tracked_physical_state(fcallback::Function, physical::T) where {T <: PhysicalState}
-    for field_symbol in fieldnames(T)
-        member = getproperty(physical, field_symbol)
-        if isa(member, TrackedVector)
-            fcallback(field_symbol, member)
-        end
-    end
-end
-
-
-"""
-Return a list of changed places in the physical state.
-A place for this state is a tuple of a symbol and the Cartesian index.
-The symbol is the name of the array within the PhysicalState.
-"""
-function changed(physical::PhysicalState)
-    places = Set{Tuple}()
-    over_tracked_physical_state(physical) do fieldname, member
-        union!(places, [(fieldname, key...) for key in changed(member)])
-    end
-    return places
-end
-
-
-"""
-Return a list of changed places in the physical state.
-A place for this state is a tuple of a symbol and the Cartesian index.
-The symbol is the name of the array within the PhysicalState.
-"""
-function wasread(physical::PhysicalState)
-    places = Set{Tuple}()
-    over_tracked_physical_state(physical) do fieldname, member
-        union!(places, [(fieldname, key...) for key in gotten(member)])
-    end
-    return places
-end
-
-"""
-Return a list of changed places in the physical state.
-A place for this state is a tuple of a symbol and the Cartesian index.
-The symbol is the name of the array within the PhysicalState.
-"""
-function resetread(physical::PhysicalState)
-    over_tracked_physical_state(physical) do _, member
-        reset_gotten!(member)
-    end
-    return physical
-end
-
-
-"""
-The arrays in a PhysicalState record that they have been modified.
-This function erases the record of modifications.
-"""
-function accept(physical::PhysicalState)
-    over_tracked_physical_state(physical) do _, member
-        reset_tracking!(member)
-    end
-    return physical
-end
-
-##### Helpers for events
-
-export EventGenerator, generators
-
-"""
-    EventGenerator{TransitionType}(matchstr, generator::Function)
-
-When an event fires, it changes the physical state. The simulation observes which
-parts of the physical state changed and sends those parts to this `EventGenerator`.
-The `EventGenerator` is a rule that matches changes to the physical state and creates
-`SimTransition` that act on that physical state.
-
-The `matchstr` is a list of symbols `(array_name, ℤ, struct_member)`. The ℤ represents
-the integer index within the array. For instance, if we simulated chess, it might
-be `(:board, ℤ, :piece)`.
-
-The generator is a callback that the simulation uses to determine which events
-need to be enabled given recent changes to the state of the board. Its signature
-is:
-
-```
-    callback_function(f::Function, physical_state, indices...)
-```
-
-Here the indices are the integer index that matches the ℤ above. This callback
-function should look at the physical state and call `f(transition)` where
-`transition` is an instance of `SimTransition`.
-"""
-struct EventGenerator{T}
-    matchstr::Vector{Symbol}
-    generator::Function
-end
-
-genmatch(eg::EventGenerator, place_key) = accessmatch(eg.matchstr, place_key)
-(eg::EventGenerator)(f::Function, physical, indices...) = eg.generator(f, physical, indices...)
-
-
-
-export EventEventGenerator
-
-
-"""
-    EventEventGenerator{TransitionType}(matchstr, generator::Function)
-
-This generator reacts to the last event fired instead of `EventGenerator` which
-reacts to the last places modified. In this case, the `matchstr` is a vector
-with one entry, the symbol version of the transition type it matches. For
-a transition called `MoveTransition` it would be `matchstr=[:MoveTransition]`.
-
-The generator is a callback function whose signature is:
-
-
-```
-    callback_function(f::Function, physical_state, event_members...)
-```
-
-This callback function is passed arguments that are the members of the instance
-of the transition it matched.
-"""
-struct EventEventGenerator
-    matchstr::Vector{Symbol}
-    generator::Function
-end
-
-genmatch(eg::EventEventGenerator, event_key) = (event_key[1] == eg.matchstr[1] ? (event_key[2:end],) : nothing)
-(eg::EventEventGenerator)(f::Function, physical, indices...) = eg.generator(f, physical, indices...)
-
-
-function transition_generate_event(gen, physical, place_key, existing_events)
-    match_result = genmatch(gen, place_key)
-    isnothing(match_result) && return nothing
-    # @debug "matched $place_key"
-    
-    # Extract the first captured integer from the ℤ⁺ pattern
-    sym_index_value = match_result[1][1]
-
-    sym_create = SimTransition[]
-    sym_depends = Set{Tuple}[]
-
-    gen(physical, sym_index_value) do transition
-        # @debug "Direction $direction"
-        resetread(physical)
-        if precondition(transition, physical)
-            input_places = wasread(physical)
-            if clock_key(transition) ∉ existing_events
-                push!(sym_create, transition)
-                push!(sym_depends, input_places)
-            end
-        end
-    end
-    if isempty(sym_create)
-        return nothing
-    else
-        return (create=sym_create, depends=sym_depends)
-    end
-end
-
-
-struct ImmediateEventGenerator
-    matchstr::Vector{Symbol}
-    generator::Function
-end
-
-genmatch(eg::ImmediateEventGenerator, event_key) = (event_key[1] == eg.matchstr[1] ? (event_key[2:end],) : nothing)
-(eg::ImmediateEventGenerator)(f::Function, physical, indices...) = eg.generator(f, physical, indices...)
-
-
-function transition_immediate_event(gen::ImmediateEventGenerator, physical, place_key, existing_events)
-    match_result = genmatch(gen, place_key)
-    isnothing(match_result) && return nothing
-    # Extract the first captured integer from the ℤ⁺ pattern
-    sym_index_value = match_result[1][1]
-
-    changed_places = Set{Tuple}()
-    gen(physical, sym_index_value) do transition
-        # @debug "Direction $direction"
-        if transition ∉ existing_events && precondition(transition, physical)
-            push!(existing_events, transition)
-            accept(physical)
-            fire!(transition, physical)
-            push!(changed_places, changed(physical))
-        end
-    end
-    return changed_places
-end
 
 ########## The Simulation Finite State Machine (FSM)
-
-"""
-  SimTransition
-
-This abstract type is the parent of all transitions in the system.
-"""
-abstract type SimTransition end
-
-"""
-InitializeEvent is a concrete transition type that represents the first event
-in the system, initialization.
-"""
-struct InitializeEvent <: SimTransition end
-
-"""
-    clock_key(::SimTransition)::Tuple
-
-All `SimTransition` objects are immutable structs that represent events but
-don't carry any mutable state. A clock key is a tuple version of an event.
-"""
-@generated function clock_key(transition::T) where T <: SimTransition
-    type_symbol = QuoteNode(nameof(T))
-    field_exprs = [:(transition.$field) for field in fieldnames(T)]
-    return :($type_symbol, $(field_exprs...))
-end
-
-"""
-    key_clock(key::Tuple)::SimTransition
-
-Takes a tuple of the form (:symbol, arg, arg) and returns an instantiation
-of the struct named by :symbol.
-"""
-@generated function key_clock(key::Tuple)
-    type_symbol = key.parameters[1]
-    if isa(type_symbol, Symbol)
-        struct_type = eval(type_symbol)
-        field_count = fieldcount(struct_type)
-        field_exprs = [:(key[$(i+1)]) for i in 1:field_count]
-        return :($struct_type($(field_exprs...)))
-    else
-        return :(error("First element of tuple must be a Symbol"))
-    end
-end
-
-
-"""
-    generators(::Type{SimTransition})::Vector{Union{EventGenerator,EventEventGenerator}}
-
-Every transition in the simulation needs generators that notice changes to state
-or events fired and create the appropriate transitions. Implement a `generators`
-function as part of the interface of each transition.
-"""
-generators(::Type{SimTransition}) = Union{EventGenerator,EventEventGenerator}[]
-
 
 mutable struct SimulationFSM{State,Sampler,CK}
     physical::State
@@ -351,19 +95,6 @@ function checksim(sim::SimulationFSM)
     @assert keys(sim.enabled_events) == keys(sim.depnet.event)
 end
 
-function capture_state_changes(f::Function, physical)
-    accept(physical)
-    result = f()
-    changes = changed(physical)
-    return (;result, changes)
-end
-
-function capture_state_reads(f::Function, physical)
-    resetread(physical)
-    result = f()
-    reads = wasread(physical)
-    return (;result, reads)
-end
 
 function rate_reenable(sim::SimulationFSM, event, clock_key)
     first_enable = sim.enabling_times[clock_key]
